renaming config files to reference singleplayer usage
