+++ apcpp-singleplayer-config.cpp
@@ -1,0 +1,5 @@
+#include <apcpp-singleplayer-config.h>
+
+namespace singleplayer {
+    
+};
